api/requirements.txt
Flask
flask-cors
Flask-JWT-Extended
geopy
psycopg2-binary
python-dotenv
gunicorn
